Create AppOperationHelpers to handle app-scoped operation-related things
